requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
numpy
pandas
paho-mqtt
joblib
scikit-learn
xgboost
python-dotenv
imbalanced-learn
requests
